Add a very simple caching system
We don't want to show parallelism of accessing the Flickr API, because this
is not the sort of thing which you want to use hardware parallelism for; if
you wanted requests dispatched in parallel, dispatch them in parallel: you
wouldn't want it to be tied to the hardware parallelism available.  So we
eliminate this source of confusion by caching responses from Flickr.  Later
we will scale images returned using this API.