Add support for TBB
The main thing to ensure is that the worker threads won't try and start the
application; here the trick of using the $threads::tbb::worker global to
check to see if it is a worker or not is used.  It would be faster (the
first time, anyway) and use less memory if the parallel code to be used was
cut down and restricted to a module.  But for this test, simplicity of
implementation is more important.