Changes for running under morbo
morbo is a restarting launcher for Mojolicious applications.  However,
using it means that '$0' is no longer the name of the script.  There's also
no sensible equivalent for FindBin in this situation, so... prefer using
the current working directory if it looks sane.

This change also demonstrates making the worker threads' module set minimal
by moving the parallel sections into an independent module file.  This
makes the worker threads more efficient.